pyproject.toml
[project]
name = "akkizon"
version = "0.1.0"
description = ""
authors = [
    {name = "Meet Ai Coders"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "openai (>=1.65.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "fastapi (>=0.115.9,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
